# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: SMIDIV API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
# basic authentication definition
securityDefinitions:
  basicAuth:
    type: basic
  tokenAuth:
    type: apiKey
    in: header
    name: X-API-KEY
paths:
...
  /brand:
    x-swagger-router-controller: marca
    get:
      security:
        - basicAuth: []
        - tokenAuth: []
      description: gets the marcaList
      operationId: getMarcaList
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      # require basic authentication on this route
      security:
        - basicAuth: []
        - tokenAuth: []
      description: Adds a new marca
      operationId: addMarca
      parameters:
        - name: marca
          description: New Marca's information
          in: body
          required: true
          schema:
            $ref: "#/definitions/Marca"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /vehicle:
    # our controller name
    x-swagger-router-controller: vehiculo
    get:
      # require basic authentication on this route
      security:
        - basicAuth: []
        - tokenAuth: []
      description: Gets the vehicles list
      operationId: getVehiclesList
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      # require basic authentication on this route
      security:
        - basicAuth: []
        - tokenAuth: []
      description: adds a new Vehiculo
      operationId: addVehiculo
      parameters:
        - name: vehiculo
          description: New Vehiculo definition
          in: body
          required: true
          schema:
            $ref: "#/definitions/Vehiculo"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Marca:
    type: object
    properties:
      nombre: 
        type: string
  Vehiculo:
    type: object
    properties:
      username:
        type: string
        description: vehicle's owner
      marca:
        type: string
        description: vehicle's brand
      modelo:
        type: string
        description: vehicle's model
      smidivID:
        type: string
      placas:
        type: string
    required:
      - username
...
  OkResponse:
    required:
      - response
      - success
    properties:
      response:
        type: object
        properties: {}
      success:
        type: boolean
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string